add authentication

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 import asyncHandler from 'express-async-handler';
 import Goal from '../models/goalModel.js';
+import User from '../models/userModel.js';
 
 // @desc Get goals
 // @route GET /api/goals
 // @access Private
 export const getGoals = asyncHandler(async (req, res) => {
-	const goals = await Goal.find({});
+	const goals = await Goal.find({ user: req.user.id });
 	res.status(200).json(goals);
 });
 
@@ -20,6 +21,7 @@ export const setGoal = asyncHandler(async (req, res) => {
 
 	const goal = await Goal.create({
 		text: req.body.text,
+		user: req.user.id,
 	});
 
 	res.status(200).json(goal);
@@ -34,6 +36,19 @@ export const updateGoal = asyncHandler(async (req, res) => {
 		res.status(404);
 		throw new Error('Goal not found');
 	}
+
+	const user = await User.findById(req.user.id);
+	// check for user
+	if (!user) {
+		res.status(401);
+		throw new Error('User not found');
+	}
+	// make sure the logged in user matches the goal user
+	if (goal.user.toString() !== user.id) {
+		res.status(401);
+		throw new Error('User not authorized');
+	}
+
 	const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
 		new: true,
 	});
@@ -49,6 +64,20 @@ export const deleteGoal = asyncHandler(async (req, res) => {
 		res.status(404);
 		throw new Error('Goal not found');
 	}
+
+	const user = await User.findById(req.user.id);
+	// check for user
+	if (!user) {
+		res.status(401);
+		throw new Error('User not found');
+	}
+
+	// make sure the logged in user matches the goal user
+	if (goal.user.toString() !== user.id) {
+		res.status(401);
+		throw new Error('User not authorized');
+	}
+
 	await Goal.findByIdAndDelete(req.params.id);
 	res.status(200).json({ id: req.params.id });
 });
